--- pages/api/hello.js
@@ -1,5 +1,0 @@
-// Next.js API route support: https://nextjs.org/docs/api-routes/introduction
-
-export default function helloAPI(req, res) {
-  res.status(200).json({ name: 'John Doe' })
-}
