--- pages/components/Header/index.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Header = () => {
-  return (
-    <div className="w-full h-24 flex justify-center items-center sticky top-0 bg-gray-100 shadow-md z-50">
-      <h4>Header da aplicaçao queria que fosse sticky</h4>
-    </div>
-  );
-};
-
-export default Header;
